Widgets: Rename PaneEmptyWidget to EmptyMessageWidget
Also, move it to its own module because it's too big to be in
helperwidgets now.
