+++ spyder/widgets/dependencies.py
@@ -24,7 +24,7 @@ from spyder.dependencies import OPTIONAL, PLUGIN
 from spyder.utils.icon_manager import ima
 from spyder.utils.palette import SpyderPalette
 from spyder.utils.stylesheet import AppStyle, MAC, WIN
-from spyder.widgets.helperwidgets import PaneEmptyWidget
+from spyder.widgets.emptymessage import EmptyMessageWidget
 
 
 class DependenciesTreeWidget(QTreeWidget):
@@ -141,7 +141,7 @@ class DependenciesDialog(QDialog):
         self.stacked_widget = QStackedWidget()
 
         # Create a loading message
-        self.loading_pane = PaneEmptyWidget(
+        self.loading_pane = EmptyMessageWidget(
             self,
             "dependencies",
             _("Dependency information will be retrieved shortly. "
